The Maximum Sum
You are given an array of integers 'a'. 
A range sum query is defined by a pair of non-negative integers 'l' and 'r' (l <= r). 
The output to a range sum query on the given array 'a' is the sum of all the elements of 'a' that have indices from 'l' to 'r', inclusive.

You have the array 'a' and a list of range sum queries 'q'. Find an algorithm that can rearrange the array 'a' in such a way that the total sum of all of the query outputs is maximized, and return this total sum.