The Paint Houses
In San Francisco, there is a row of several beautiful houses called the Painted Ladies. 
Each of the Painted Ladies can be painted with one of three colors: red, blue or green. 
The cost of painting each house with a certain color is different. cost[i][0] for each 'i' is the cost of painting house 'i' red, cost[i][1] is the cost of painting it blue, and cost[i][2] is the cost of painting it green.

You want to paint all the houses in a way such that no two adjacent Painted Ladies have the same color. 
Find the minimum cost to achieve this.