The Knight On Board Probability
You have a chessboard with 'n' rows and 'm' columns. 
The knight is placed on the (x, y) cell (0-indexed). 
Find the probability that the knight stays on the board after performing the usual knight move steps times randomly. 
If the knight gets moved off of the board, it can't return to the board again.