The Equilibrium Point
Equilibrium position in an array is a position at which the sum of elements before it is equal to the sum of elements after it. 
Given an array 'a', your task is to determine at which position equilibrium first occurs in the array. 
If there is no equilibrium position, the answer should be -1.