The Find Subarray By Sum
You're given an unsorted array 'a' of positive integers and a number 's'. 
Your task is to find a contiguous subarray that has a sum equal to 's', and return an array containing the two integers that represent its inclusive bounds. 
If there are several possible answers, return the one with the smallest left bound. 
If there are no answers, return [-1].
Your answer should be 1-based, making the first position of the array 1 instead of 0.