The Rate Limit
Your company decides to give Kik a try and write a chat bot to be used on the platform. You've worked hard and it's almost finished, and all that's left to do now is test it. One of the things you want to check is whether the number of messages your bot sends exceeds the per user rate limit.

The purpose of the per user rate limit is to prevent bots from sending an excessively high number of messages to individual users every day. This limit is defined on a per-user basis: at midnight each day (00:00 in local server time, which also happens to be Coordinated Universal Time) the baseline limit is reset to 'startingAllowance'. Once your bot sends a message to a given user that day, the per user rate limit (for that user) is decreased by 1. Whenever a user messages the bot, the per user rate limit increases by 1 (for that user). If at some point your bot tries to send a message to a user whose current per user rate limit equals 0, the sending fails.

Importantly, the rate limit is applied to batches of messages sent to multiple users. For example, if a batch of 25 messages is sent to various users, but the rate limit of at least one of them is 0, then the whole batch fails and no message is sent to any of the users.

You're given the logs of the sentBatches of messages your bot sent without setting the per user rate limit, and information about the 'receivedMessages'. Which of the sent batches would fail if there was a per user rate limit with the initial value of 'startingAllowance'?