The Data Route
Two Sigma's mission is to uncover value in the world's data, and as part of that it's necessary to download massive amounts of information on a regular basis. Naturally, this data should be transferred as quickly and efficiently as possible.

A new data resource was recently added to the network, and your job is to establish a connection between it and Two Sigma's server. Due to security reasons, all connections in Two Sigma's network are unidirectional (i.e. only have a one-way connection), and no data can be stored on any node of the network. This means that every second the amount of data a node receives should be equal to the amount of data it forwards. The only exceptions to this rule are resource and server, since the former only sends data while the latter only receives it.

Unfortunately, some segments of the network are slower than is ideal due to limitations with legacy telecommunication operators around the world. This complicates finding the optimal route through the network significantly, which is why your help is required.

Find a route between the resource and the server that maximizes the amount of data downloaded in a second, and return this maximum value.