The Concurrent Backups
Datto is designed to perform backups as quickly as possible, which is usually achieved by using multiple threads. In this task, assume that all of your available threads are backing up documents at the same 1Mb/sec speed.

Given an array of documents sizes that you need to back up and the number of available threads, return the minimum amount of time needed to back up all the files (a single thread can only be backing up one document at any given moment).