The Busy Holidays
To celebrate Cyber Monday, Instacart has decided to give its shoppers (employees that shop at grocery stores and deliver orders to customers) a break. For a 24h period, every shopper only has to make 1 delivery, after which his/her workday is over. However, since providing customers with a reliable shopping experience is Instacart's #1 priority, it is important to ensure that each order is fulfilled and delivered as promised.

You are given a list of orders with the time periods when they should be completed, and the time leadTime it takes to fulfill each order. Knowing the time period each shopper is available (shoppers), find out whether the current number of shoppers is enough to fulfill all orders.

A shopper can fulfill an order if he/she can both start and finish it in the time period specified for this order.