The Kik Code
Kik Codes are the fastest, easiest way to add friends and join groups on Kik. Just scan the code with your device, and a chat with the corresponding user starts automatically.

As a part of your preparation for an interview at Kik, you decide to implement an algorithm that will automatically build a Kik Code given a userId. As you can see, a code consists of 6 circles with their circumferences partly colored. You're going to take the binary representation of userId, extend it by leading zeros if necessary (there should be 52 bits in its binary representation), reverse it, and then map it to the Code as follows:

CIRCUMFERENCE	1	      2	  3	      4	    5	         6
BITS	                        first 3     next 4  next 8  next 10  next 12	 last 15

Each circumference will be split into sectors. The number of sectors for each circumference is equal to the number of bits mapped onto it. The contiguous sets of 1s in the corresponding bits of 'userId' will be colored white.

Now you need to implement a function that, given userId, returns the list of colored segments' coordinates in the 'polar coordinate system' with the center at the center of the circumferences. Each segment should be returned as an array of two elements, representing the coordinates of its edges in a counterclockwise direction.