The Launch Sequence Checker
The master launch sequence consists of several independent sequences for different systems. Your goal is to verify that all the individual system sequences are in strictly increasing order. In other words, for any two elements i and j (i < j) of the master launch sequence that belong to the same system (having a[i] = a[j]), their values should be in strictly increasing order (i.e. b[i] < b[j]).