The Red Alert
After long years of sharpening your programming skills and honing your physique you finally made it: as a member of the ProProgrammers team you've entered 'Fort Boyard'!

Your team has successfully completed several challenges so far, and one of your teammates is ready to face the next one. In this challenge the participant should make her way through a corridor covered in red laser rays. If she touches one of the rays, she will fail the challenge and become the prisoner of the Fort.

The corridor has several levels, so it can be represented as an undirected graph of the known number of nodes with start at node 1 and finish at node nodes, both 1-based. Each move the participant can move to the node adjacent to her current position. Since you prepared for each challenge thoroughly, you know when and where the laser rays appear: on the ith (1-based) move the rays disappear from their previous locations and appear in all the nodes x, such that x % k = i % k, where % stands for the modulo operation and k is known. If your teammate ends up in node x when the rays appear there, she'll fail.

Your goal is to calculate the minimum number of moves required to reach the last node from the node 1, or -1 if it's impossible.