The Job Scheduling
Two Sigma maintains a distributed computing cluster capable of handling many different types of jobs. Some jobs are quite small (for instance MapReduce style jobs), but others can be larger (certain types of distributed optimization).

Let's assume that the jobs are processed one by one, and if a new job request comes in when one is already being processed it's added to the queue. You were asked to implement a scheduling algorithm that adds jobs to the regular queue and pushes them through in such a way that the average wait time for all jobs in the queue is minimized. A new job isn't pushed through unless it minimizes the average waiting time.

Assume that your program starts working at 0 seconds. A request for the ith job came at 'requestTime(rT)', and let's assume that it takes 'jobProcess(jP)' seconds to process it.

Find the state of your algorithm's queue 'timeFromStart(tFS)' seconds after your program started to work, assuming that all actions that could've happened right at this moment have already happened.