The Road Map
You have a roadmap, which is the list of tasks (t) that your team needs to complete. Each task in this list has a title, a start date (st), an end date (ft), and a list of the people who will be working on it. You are given some queries (q), each of which contains a specific person's name and a date. For each query that is made, you need to return the list of tasks on which that person will be working on the date specified in the query, sorted by the tasks' end dates. If their end dates are equal, then sort by the tasks' titles.