The Army March
Picture a map of a battlefield divided into two halves. Points which lie to the left of the line x = 0 are on the forest free side, while points which lie on or to the right of the line are in the forest.

Your army needs to get from point a to point b. The problem is that while point a is located on the forest free side where your army's speed equals v1 miles per hour, point b is located in the forest, where your speed reduces to v2 miles per hour.

Calculate the shortest possible travel time for your army to reach point b.