The Opponent Matching
When you click the VS Fight button on CodeSignal, the system tries to match you with the best opponent possible. The matching algorithm has become more complex over time,but initially it was a simple search for someone whose xp is as close to yours as possible.

The easiest way to understand how it used to conduct the search is as follows:
  - Imagine that each user looking for an opponent is standing at the center of a search circle on a horizontal xp axis.
  - All the search circles have the same radius (the search radius), and initially search radius is equal to 0.
  - At each step, the search radius is increased by 1.
  - A match is found as soon as two search circles intersect. These circles are then removed immediately.
  - For the sake of simplicity, assume that on each step no more than one pair of circles can intersect.
Given a list of requests as user xps, match them up using the algorithm described above.