The Great Renaming
Once upon a time, in a kingdom far, far away, there lived a King Byteasar VI. As any king with such a magnificent name, he was destined to leave a trace in history. Unfortunately, imagination wasn't one of Byteasar's strong suits, so the reform he came up with was quite simple: he ordered that all his kingdom's cities should be renamed. He didn't want to come up with new names (as a king, he'd have to remember them all!), so he ordered the cities to swap their names in such a manner that the ith city would have the name of the city number (i + 1)th. The last city in the Byteasar's kingdom would, naturally, get the name of the very first city.

The cartographers were not happy with this reform, since they had to redraw all the kingdom's maps. Before the reform, information about all the roads between the cities were stored in the 'a'. Prior to redrawing any map, they had to start with updating the register. Your task is to find out the state of the updated register.