The Caterpillar Trees
Not long ago you discovered a wonderful tree in the nearby woods that made you very curious about the greenery around you. You went to the nearby woods and drew various plants in your notebook. The plants in the woods have various structures: some of them even contain loops!

Anyway, for now, you are interested only in trees. You came up with a brand new property: you call a tree a caterpillar if there exists a path in it, such that each vertex of a tree either belongs to this path or is connected to one of its vertices by a single edge. To find out more about them, you would like to write a program that will find all the interesting trees in the structures you drew in your notebook.

The plants you drew consist of n vertices and are connected by several edges. Calculate the number of regular trees and caterpillar trees in the structures you drew in your notebook.